Change the Rust WebAssembly compilation target
Use the Rust target `wasm32-wasi` instead of the `wasm32-unknown-unknown`.
This is required to prevent a Rust panic to happen when the policy is loaded.

For more details, see https://github.com/kubewarden/rust-policy-template/pull/25

On top of that, the Makefile rule has been changed to use `cp` instead of `mv`.
Without this change, invoking the `make e2e-tests` twice in a row, caused errors
when no change was done to the source code of the policy.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>